requirements.txt
# PyTorch with CUDA 11.8 (RTX 3060 optimized)
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Data processing
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
ta==0.11.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.20

# Backend API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.66.1
matplotlib==3.8.0
seaborn==0.13.0

# Already using
yfinance==0.2.28
